+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed invocation of server functions that do not return `Promise` instances by calling them in the handler of a resolved promise.
 
 ## [0.77.2] - 2021-02-27
 ### Fixed